Work on upgrading key manager into a full capability system

+++ kernel/managers/syscall_manager/syscalls/priv_syscalls.h
@@ -66,11 +66,11 @@ uint64_t GetSharedMemoryKey_Syscall(uint64_t vAddress,
                                     uint64_t length,
                                     CachingMode cacheMode,
                                     MemoryAllocationFlags flags,
-                                    uint128_t *key);
+                                    uint8_t *key);
 
-uint64_t ApplySharedMemoryKey_Syscall(uint128_t* key, void* shmem_data_p);
+uint64_t ApplySharedMemoryKey_Syscall(uint8_t* key, void* shmem_data_p);
 
-uint64_t FreeSharedMemoryKey_Syscall(uint128_t* key);
+uint64_t FreeSharedMemoryKey_Syscall(uint8_t* key);
 
 uint64_t R0_AllocateSharedMemory_Syscall(uint64_t length,
                                          CachingMode cacheMode,
@@ -78,7 +78,7 @@ uint64_t R0_AllocateSharedMemory_Syscall(uint64_t length,
                                          MemoryAllocationFlags flags,
                                          uint64_t phys_addr);
 
-uint64_t GetSharedMemoryKeyUsageCount_Syscall(uint128_t *key);
+uint64_t GetSharedMemoryKeyUsageCount_Syscall(uint8_t *key);
 
 uint64_t R01_RegisterForInterrupts_Syscall(uint64_t UNUSED(instruction_pointer),
                                            uint64_t syscall_num,
